requirements.txt
Flask==3.0.0
moviepy==1.0.3
gTTS==2.5.0
Pillow==10.3.0
numpy==1.26.4
requests==2.31.0
